requirements.txt
grpcio==1.28.1
grpcio-tools==1.28.1
numpy==1.16.1
